test: add a test for the combination of Nunchaku with TeaCache
