+++ examples/flux.1-dev-teacache.py
@@ -1,407 +1,12 @@
 import time
-from typing import Any, Callable, Optional, Union
 
-import numpy as np
 import torch
-from diffusers.models.modeling_outputs import Transformer2DModelOutput
-from diffusers.models.transformers.transformer_flux import FluxTransformer2DModel
 from diffusers.pipelines.flux.pipeline_flux import FluxPipeline
-from diffusers.utils import logging
-from diffusers.utils.constants import USE_PEFT_BACKEND
-from diffusers.utils.import_utils import is_torch_version
-from diffusers.utils.peft_utils import scale_lora_layers, unscale_lora_layers
 
 from nunchaku import NunchakuFluxTransformer2dModel
 from nunchaku.utils import get_precision
+from teacache import TeaCache
 
-logger = logging.get_logger(__name__)  # pylint: disable=invalid-name
-
-def make_teacache_forward(
-    num_steps: int = 50, rel_l1_thresh: float = 0.6, skip_steps: int = 0
-) -> Callable:
-    def teacache_forward(
-        self: Union[FluxTransformer2DModel, NunchakuFluxTransformer2dModel],
-        hidden_states: torch.Tensor,
-        encoder_hidden_states: torch.Tensor,
-        pooled_projections: torch.Tensor,
-        timestep: torch.LongTensor,
-        img_ids: torch.Tensor,
-        txt_ids: torch.Tensor,
-        guidance: torch.Tensor,
-        joint_attention_kwargs: Optional[dict[str, Any]] = None,
-        controlnet_block_samples: Optional[torch.Tensor] = None,
-        controlnet_single_block_samples: Optional[torch.Tensor] = None,
-        return_dict: bool = True,
-        controlnet_blocks_repeat: bool = False,
-    ) -> Union[torch.FloatTensor, Transformer2DModelOutput]:
-        """
-        The [`FluxTransformer2DModel`] forward method.
-
-        Args:
-            hidden_states (`torch.FloatTensor` of shape `(batch size, channel, height, width)`):
-                Input `hidden_states`.
-            encoder_hidden_states (`torch.FloatTensor` of shape `(batch size, sequence_len, embed_dims)`):
-                Conditional embeddings (embeddings computed from the input conditions such as prompts) to use.
-            pooled_projections (`torch.FloatTensor` of shape `(batch_size, projection_dim)`): Embeddings projected
-                from the embeddings of input conditions.
-            timestep ( `torch.LongTensor`):
-                Used to indicate denoising step.
-            block_controlnet_hidden_states: (`list` of `torch.Tensor`):
-                A list of tensors that if specified are added to the residuals of transformer blocks.
-            joint_attention_kwargs (`dict`, *optional*):
-                A kwargs dictionary that if specified is passed along to the `AttentionProcessor` as defined under
-                `self.processor` in
-                [diffusers.models.attention_processor](https://github.com/huggingface/diffusers/blob/main/src/diffusers/models/attention_processor.py).
-            return_dict (`bool`, *optional*, defaults to `True`):
-                Whether or not to return a [`~models.transformer_2d.Transformer2DModelOutput`] instead of a plain
-                tuple.
-
-        Returns:
-            If `return_dict` is True, an [`~models.transformer_2d.Transformer2DModelOutput`] is returned, otherwise a
-            `tuple` where the first element is the sample tensor.
-        """
-        if joint_attention_kwargs is not None:
-            joint_attention_kwargs = joint_attention_kwargs.copy()
-            lora_scale = joint_attention_kwargs.pop("scale", 1.0)
-        else:
-            lora_scale = 1.0
-
-        if USE_PEFT_BACKEND:
-            # weight the lora layers by setting `lora_scale` for each PEFT layer
-            scale_lora_layers(self, lora_scale)
-        else:
-            if (
-                joint_attention_kwargs is not None
-                and joint_attention_kwargs.get("scale", None) is not None
-            ):
-                logger.warning(
-                    "Passing `scale` via `joint_attention_kwargs` when not using the PEFT backend is ineffective."
-                )
-
-        hidden_states = self.x_embedder(hidden_states)
-
-        timestep = timestep.to(hidden_states.dtype) * 1000  # type: ignore
-        if guidance is not None:
-            guidance = guidance.to(hidden_states.dtype) * 1000
-        else:
-            guidance = None
-
-        temb = (
-            self.time_text_embed(timestep, pooled_projections)
-            if guidance is None
-            else self.time_text_embed(timestep, guidance, pooled_projections)
-        )
-        encoder_hidden_states = self.context_embedder(encoder_hidden_states)
-
-        if txt_ids.ndim == 3:
-            logger.warning(
-                "Passing `txt_ids` 3d torch.Tensor is deprecated."
-                "Please remove the batch dimension and pass it as a 2d torch Tensor"
-            )
-            txt_ids = txt_ids[0]
-        if img_ids.ndim == 3:
-            logger.warning(
-                "Passing `img_ids` 3d torch.Tensor is deprecated."
-                "Please remove the batch dimension and pass it as a 2d torch Tensor"
-            )
-            img_ids = img_ids[0]
-
-        ids = torch.cat((txt_ids, img_ids), dim=0)
-        image_rotary_emb = self.pos_embed(ids)
-
-        if (
-            joint_attention_kwargs is not None
-            and "ip_adapter_image_embeds" in joint_attention_kwargs
-        ):
-            ip_adapter_image_embeds = joint_attention_kwargs.pop("ip_adapter_image_embeds")
-            ip_hidden_states = self.encoder_hid_proj(ip_adapter_image_embeds)  # type: ignore
-            joint_attention_kwargs.update({"ip_hidden_states": ip_hidden_states})
-
-        inp = hidden_states.clone()
-        temb_ = temb.clone()
-        modulated_inp, gate_msa, shift_mlp, scale_mlp, gate_mlp = self.transformer_blocks[0].norm1(inp, emb=temb_)  # type: ignore
-        if self.cnt == 0 or self.cnt == num_steps - 1:
-            should_calc = True
-            self.accumulated_rel_l1_distance = 0
-        else:
-            coefficients = [
-                4.98651651e02,
-                -2.83781631e02,
-                5.58554382e01,
-                -3.82021401e00,
-                2.64230861e-01,
-            ]
-            rescale_func = np.poly1d(coefficients)
-            self.accumulated_rel_l1_distance += rescale_func(
-                (
-                    (modulated_inp - self.previous_modulated_input).abs().mean()
-                    / self.previous_modulated_input.abs().mean()
-                )
-                .cpu()
-                .item()
-            )
-            if self.accumulated_rel_l1_distance < rel_l1_thresh:
-                should_calc = False
-            else:
-                should_calc = True
-                self.accumulated_rel_l1_distance = 0
-        self.previous_modulated_input = modulated_inp
-        self.cnt += 1
-        if self.cnt == num_steps:
-            self.cnt = 0
-
-        ckpt_kwargs: dict[str, Any]
-
-        if self.cnt > skip_steps:
-            if not should_calc:
-                hidden_states += self.previous_residual
-            else:
-                ori_hidden_states = hidden_states.clone()
-                for index_block, block in enumerate(self.transformer_blocks):
-                    if torch.is_grad_enabled() and self.gradient_checkpointing:
-
-                        def create_custom_forward(module, return_dict=None):  # type: ignore
-                            def custom_forward(*inputs):  # type: ignore
-                                if return_dict is not None:
-                                    return module(*inputs, return_dict=return_dict)
-                                else:
-                                    return module(*inputs)
-
-                            return custom_forward
-
-                        ckpt_kwargs = (
-                            {"use_reentrant": False} if is_torch_version(">=", "1.11.0") else {}
-                        )
-                        encoder_hidden_states, hidden_states = torch.utils.checkpoint.checkpoint(
-                            create_custom_forward(block),
-                            hidden_states,
-                            encoder_hidden_states,
-                            temb,
-                            image_rotary_emb,
-                            **ckpt_kwargs,
-                        )
-
-                    else:
-                        encoder_hidden_states, hidden_states = block(
-                            hidden_states=hidden_states,
-                            encoder_hidden_states=encoder_hidden_states,
-                            temb=temb,
-                            image_rotary_emb=image_rotary_emb,
-                            joint_attention_kwargs=joint_attention_kwargs,
-                        )
-
-                    # controlnet residual
-                    if controlnet_block_samples is not None:
-                        interval_control = len(self.transformer_blocks) / len(
-                            controlnet_block_samples
-                        )
-                        interval_control = int(np.ceil(interval_control))
-                        # For Xlabs ControlNet.
-                        if controlnet_blocks_repeat:
-                            hidden_states = (
-                                hidden_states
-                                + controlnet_block_samples[
-                                    index_block % len(controlnet_block_samples)
-                                ]
-                            )
-                        else:
-                            hidden_states = (
-                                hidden_states
-                                + controlnet_block_samples[index_block // interval_control]
-                            )
-                hidden_states = torch.cat([encoder_hidden_states, hidden_states], dim=1)
-
-                for index_block, block in enumerate(self.single_transformer_blocks):
-                    if torch.is_grad_enabled() and self.gradient_checkpointing:
-
-                        def create_custom_forward(module, return_dict=None):  # type: ignore
-                            def custom_forward(*inputs):  # type: ignore
-                                if return_dict is not None:
-                                    return module(*inputs, return_dict=return_dict)
-                                else:
-                                    return module(*inputs)
-
-                            return custom_forward
-
-                        ckpt_kwargs = (
-                            {"use_reentrant": False} if is_torch_version(">=", "1.11.0") else {}
-                        )
-                        hidden_states = torch.utils.checkpoint.checkpoint(
-                            create_custom_forward(block),
-                            hidden_states,
-                            temb,
-                            image_rotary_emb,
-                            **ckpt_kwargs,
-                        )
-
-                    else:
-                        hidden_states = block(
-                            hidden_states=hidden_states,
-                            temb=temb,
-                            image_rotary_emb=image_rotary_emb,
-                            joint_attention_kwargs=joint_attention_kwargs,
-                        )
-
-                    # controlnet residual
-                    if controlnet_single_block_samples is not None:
-                        interval_control = len(self.single_transformer_blocks) / len(
-                            controlnet_single_block_samples
-                        )
-                        interval_control = int(np.ceil(interval_control))
-                        hidden_states[:, encoder_hidden_states.shape[1] :, ...] = (
-                            hidden_states[:, encoder_hidden_states.shape[1] :, ...]
-                            + controlnet_single_block_samples[index_block // interval_control]
-                        )
-
-                hidden_states = hidden_states[:, encoder_hidden_states.shape[1] :, ...]
-                self.previous_residual = hidden_states - ori_hidden_states
-        else:
-            for index_block, block in enumerate(self.transformer_blocks):
-                if torch.is_grad_enabled() and self.gradient_checkpointing:
-
-                    def create_custom_forward(module, return_dict=None):  # type: ignore
-                        def custom_forward(*inputs):  # type: ignore
-                            if return_dict is not None:
-                                return module(*inputs, return_dict=return_dict)
-                            else:
-                                return module(*inputs)
-
-                        return custom_forward
-
-                    ckpt_kwargs = (
-                        {"use_reentrant": False} if is_torch_version(">=", "1.11.0") else {}
-                    )
-                    encoder_hidden_states, hidden_states = torch.utils.checkpoint.checkpoint(
-                        create_custom_forward(block),
-                        hidden_states,
-                        encoder_hidden_states,
-                        temb,
-                        image_rotary_emb,
-                        **ckpt_kwargs,
-                    )
-
-                else:
-                    encoder_hidden_states, hidden_states = block(
-                        hidden_states=hidden_states,
-                        encoder_hidden_states=encoder_hidden_states,
-                        temb=temb,
-                        image_rotary_emb=image_rotary_emb,
-                        joint_attention_kwargs=joint_attention_kwargs,
-                    )
-
-                # controlnet residual
-                if controlnet_block_samples is not None:
-                    interval_control = len(self.transformer_blocks) / len(controlnet_block_samples)
-                    interval_control = int(np.ceil(interval_control))
-                    # For Xlabs ControlNet.
-                    if controlnet_blocks_repeat:
-                        hidden_states = (
-                            hidden_states
-                            + controlnet_block_samples[index_block % len(controlnet_block_samples)]
-                        )
-                    else:
-                        hidden_states = (
-                            hidden_states
-                            + controlnet_block_samples[index_block // interval_control]
-                        )
-            hidden_states = torch.cat([encoder_hidden_states, hidden_states], dim=1)
-
-            for index_block, block in enumerate(self.single_transformer_blocks):
-                if torch.is_grad_enabled() and self.gradient_checkpointing:
-
-                    def create_custom_forward(module, return_dict=None):  # type: ignore
-                        def custom_forward(*inputs):  # type: ignore
-                            if return_dict is not None:
-                                return module(*inputs, return_dict=return_dict)
-                            else:
-                                return module(*inputs)
-
-                        return custom_forward
-
-                    ckpt_kwargs = (
-                        {"use_reentrant": False} if is_torch_version(">=", "1.11.0") else {}
-                    )
-                    hidden_states = torch.utils.checkpoint.checkpoint(
-                        create_custom_forward(block),
-                        hidden_states,
-                        temb,
-                        image_rotary_emb,
-                        **ckpt_kwargs,
-                    )
-
-                else:
-                    hidden_states = block(
-                        hidden_states=hidden_states,
-                        temb=temb,
-                        image_rotary_emb=image_rotary_emb,
-                        joint_attention_kwargs=joint_attention_kwargs,
-                    )
-
-                # controlnet residual
-                if controlnet_single_block_samples is not None:
-                    interval_control = len(self.single_transformer_blocks) / len(
-                        controlnet_single_block_samples
-                    )
-                    interval_control = int(np.ceil(interval_control))
-                    hidden_states[:, encoder_hidden_states.shape[1] :, ...] = (
-                        hidden_states[:, encoder_hidden_states.shape[1] :, ...]
-                        + controlnet_single_block_samples[index_block // interval_control]
-                    )
-
-            hidden_states = hidden_states[:, encoder_hidden_states.shape[1] :, ...]
-
-        hidden_states = self.norm_out(hidden_states, temb)
-        output: torch.FloatTensor = self.proj_out(hidden_states)
-
-        if USE_PEFT_BACKEND:
-            # remove `lora_scale` from each PEFT layer
-            unscale_lora_layers(self, lora_scale)
-
-        if not return_dict:
-            return output
-
-        return Transformer2DModelOutput(sample=output)
-
-    return teacache_forward
-
-
-# A context manager to add teacache support to a block of code
-# When the context manager is applied, the model passed to the context manager is modified
-# to support teacache
-class TeaCache:
-    def __init__(
-        self,
-        model: Union[FluxTransformer2DModel, NunchakuFluxTransformer2dModel],
-        num_steps: int = 50,
-        rel_l1_thresh: float = 0.6,
-        skip_steps: int = 0,
-        enabled: bool = True,
-    ) -> None:
-        self.model = model
-        self.num_steps = num_steps
-        self.rel_l1_thresh = rel_l1_thresh
-        self.skip_steps = skip_steps
-        self.enabled = enabled
-        self.previous_model_forward = self.model.forward
-
-    def __enter__(self) -> "TeaCache":
-        if self.enabled:
-            self.model.__class__.forward = make_teacache_forward(self.num_steps, self.rel_l1_thresh, self.skip_steps)  # type: ignore
-            self.model.cnt = 0
-            self.model.accumulated_rel_l1_distance = 0
-            self.model.previous_modulated_input = None
-            self.model.previous_residual = None
-        return self
-
-    def __exit__(self, exc_type: Any, exc_value: Any, traceback: Any) -> None:
-        if self.enabled:
-            self.model.__class__.forward = self.previous_model_forward  # type: ignore
-            del self.model.cnt
-            del self.model.accumulated_rel_l1_distance
-            del self.model.previous_modulated_input
-            del self.model.previous_residual
-
-# Actual inference starts here
 
 precision = get_precision()  # auto-detect your precision is 'int4' or 'fp4' based on your GPU
 transformer = NunchakuFluxTransformer2dModel.from_pretrained(f"mit-han-lab/svdq-{precision}-flux.1-dev")
@@ -425,4 +30,4 @@ with TeaCache(
     ).images[0]
 end_time = time.time()
 print(f"Time taken: {(end_time - start_time)} seconds")
-image.save(f"flux.1-dev-{precision}.png")
+image.save(f"flux.1-dev-{precision}-tc.png")
